refactor: decompose meatspace.js router into per-domain sub-routers + thin barrel (#1833)
